test: Moved group show test into groups
